+++ ios/Classes/FluttersliderdrawerPlugin.h
@@ -1,4 +1,0 @@
-#import <Flutter/Flutter.h>
-
-@interface FluttersliderdrawerPlugin : NSObject<FlutterPlugin>
-@end